+++ src/app/content/highlights/components/SummaryPopup/Filters.tsx
@@ -2,10 +2,10 @@ import flow from 'lodash/fp/flow';
 import React from 'react';
 import { connect } from 'react-redux';
 import { AppState, Dispatch } from '../../../../types';
-import ChapterFilter from '../../../elements/popUp/ChapterFilter';
-import Filters, { FilterDropdown } from '../../../elements/popUp/Filters';
-import FiltersList from '../../../elements/popUp/FiltersList';
-import PrintButton from '../../../elements/popUp/PrintButton';
+import ChapterFilter from '../../../components/popUp/ChapterFilter';
+import Filters, { FilterDropdown } from '../../../components/popUp/Filters';
+import FiltersList from '../../../components/popUp/FiltersList';
+import PrintButton from '../../../components/popUp/PrintButton';
 import { printSummaryHighlights, setSummaryFilters } from '../../actions';
 import * as select from '../../selectors';
 import ColorFilter from './ColorFilter';
